# Add support for passing Configs to Dataclasses