<!-- Thank you for your contribution! -->
<!-- Don't worry about making lots of commits on a pull request, they'll be squashed on merge anyway -->

## Change Summary

Enabled support for passing Config classes to models based on `@dataclass` decorator.

## Related issue number

#276

## Performance Changes

pydantic cares about performance, if there's any risk performance changed on this PR, 
please run `make benchmark-pydantic` before and after the change:
* before: **pydantic time=1.276s, success=47.10% pydantic best=1.276s, avg=1.276s, stdev=0.000s**
* after: **1.204s, success=48.55% pydantic best=1.204s, avg=1.204s, stdev=0.000s**

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes
* [x] No performance deterioration (if applicable)
* [x] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * if you're not a regular contributer please include your github username `@whatever`