@samuelcolvin 2018-11-15
Hi @jarekkar would love to get this merged if you had time to work on it?

Slightly annoyingly, I've merged the black PR #287 so you'll probably have a painful time rebasing. If it's too annoying, let me know and I'll do it.