@jarekkar 2018-11-15
Hi, currently I am on the business trip and I don't have time. I'll be back
at home on beginning of the next week. And then I'll try. You don't need to
wait for me. I will rebase the project.
--
Pozdrawiam
Jarek


czw., 15 lis 2018 o 12:39 Samuel Colvin <notifications@github.com>
napisał(a):

> Hi @jarekkar <https://github.com/jarekkar> would love to get this merged
> if you had time to work on it?
>
> Slightly annoyingly, I've merged the black PR #287
> <https://github.com/samuelcolvin/pydantic/pull/287> so you'll probably
> have a painful time rebasing. If it's too annoying, let me know and I'll do
> it.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/samuelcolvin/pydantic/pull/285#issuecomment-439011789>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AO5WZkyJLP58k5PEG7rI2OBPVq-cg-A_ks5uvVKIgaJpZM4X9tqB>
> .
>